Describe credentials after the session creation
The `describe-creds` was invoked after the `session-ic` was run.
However, it always showed `Cannot find AWS credentials configured by
'aws-creds'.` as the environment variables were not updated before
describing the connection.

The new version does not rely on the environment variables to describe
credentials during the session creation; instead, they are filled out
based on the session information.