Switch from positional to named command arguments
`aws-creds` versions <= `0.4.x` used the positional arguments to pass
the parameters. The positional ones are excellent for the constant
number of parameters and introduce challenges when you need a variable
list of parameters. Moreover, it complicates further development of
features requiring optional inputs, such as configuring the
`AWS_DEFAULT_OUTPUT` format for every alias. That’s why the named
parameters are used instead now. To maintain backward compatibility, the
`session-ic` command will continue to process the positional arguments
by prefixing them with a deprecation message.