Prefix aliases with `aws-creds`
Based on the feedback, running an alias in CLI that is equal to the AWS
profile name can be confusing. Although the alias name could be updated
manually, the initial aliases will now be prefixed with `aws-creds-`.
This change provides a common namespace for all the `aws-creds`-related
commands and built-in shell autocompletion.