Use `~/.local/bin` for Test Installations
The `~/.local/bin` directory is preferred for test installations as it
resides within the user's HOME directory and does not require `sudo`
permissions for installing scripts. In contrast, `/usr/local/bin` may
require elevated permissions, which could complicate the process.

For this reason, the installation location has been moved to
`~/.local/bin`.