Control the browser behavior.
AWS IAM Identity Center asks to verify the device when creating a new
session. Before, `aws-creds` opened the verification URL automatically
in a default browser. However, it does not work for an environment
without a browser or for an environment with an insufficient default
browser. That’s why the new logic allows controlling the browser's
behavior as

1. `scan-ic` and `scan-local` commands ask the user if the verification
   URL should be open automatically.
2. `session-ic` opens the URL by default. However, the `--no-browser`
   option lets you turn it off.